ECOPROJECT-3905 | fix: sort clusters by VM count descending and auto-select the largest cluster
1. Move cluster-selection business logic out of views layer
2. Validated persisted manual selection
3. Extracted VM-count comparator to domain layer
